Create README.md
This repo contains patches needed to build AOSP using Hisense's proprietary blobs.  It's layout parallels that of the individual git repos in the AOSP tree.  Except for the 'device' subtree, each patch can be applied directly to the corresponding repo.

The patches under 'device' contain mods that other developers may want to apply to their own 'device' and 'vendor' trees.  They should not be applied to my 'm470bsa' repos.